+++ src/client/java/minicraft/level/tile/HardRockTile.java
@@ -65,14 +65,14 @@ public class HardRockTile extends Tile {
 		int damage = level.getData(x, y) + dmg;
 		int hrHealth = 200;
 		if (Game.isMode("minicraft.settings.mode.creative")) dmg = damage = hrHealth;
-		level.add(new SmashParticle(x * 16, y * 16));
+		level.add(new SmashParticle(x << 4, y << 4));
 		Sound.play("monsterhurt");
 
-		level.add(new TextParticle("" + dmg, x * 16 + 8, y * 16 + 8, Color.RED));
+		level.add(new TextParticle("" + dmg, (x << 4) + 8, (y << 4) + 8, Color.RED));
 		if (damage >= hrHealth) {
 			level.setTile(x, y, Tiles.get("dirt"));
-			level.dropItem(x * 16 + 8, y * 16 + 8, 1, 3, Items.get("Stone"));
-			level.dropItem(x * 16 + 8, y * 16 + 8, 0, 1, Items.get("Coal"));
+			level.dropItem((x << 4) + 8, (y << 4) + 8, 1, 3, Items.get("Stone"));
+			level.dropItem((x << 4) + 8, (y << 4) + 8, 0, 1, Items.get("Coal"));
 		} else {
 			level.setData(x, y, damage);
 		}